Added Database with content provider.
View nodes fragment listens for changes and updates in the app database and updates changes in the list of nodes automatically.

In database, following details of a node are being saved:
1. Name - Same as SSID, USERDEFINED
2. SSID
3. Level - PRIMARY, SECONDARY, VITAL
4. Location - NEW, USERDEFINED
5. Type	- Light, Fan, etc
6. State - ON, OFF

The nodes are to be assigned one of the three levels:
1. PRIMARY
2. SECONDARY
3. VITAL

SECONDARY level is assigned by default.

Applciation has 3 fragments:
1. View all nodes
2. Master switch
3. Add new nodes

View all nodes has following capabilities:
1. Swipe to refresh
2. Toggle state of individual nodes

Master switch toggles state for all the nodes at the level: PRIMARY.

Updates to be expected in near future:
1. A details activity will soon be introduced through which the user can edit the existing node.
2. Add new node fragment completion
3. Login
4. Swipe to remove node in the View all nodes fragment.